requirements.txt
streamlit==1.35.0
ollama==0.1.9
python-dotenv==1.0.1
google-generativeai==0.4.0
watchdog==4.0.0
